feat: create Status bar component
