+++ src/components/molecules/StatusBar/style.module.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const StatusBarBody = styled.div`
+  margin: 25px;
+`
